Added StaticString and StaticArray. replaced all int types with stdint.h ones.

+++ Source/Code/ArduinoTemplateLibrary/Arduino/StreamInputStream.h
@@ -58,7 +58,7 @@ public:
 	/*
 		Returns the number of bytes that are available in the stream.
 	 */
-	unsigned int getLength() const
+	uint16_t getLength() const
 	{
 		if (_stream == NULL) return 0;
